@@ -10,23 +10,23 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.0.3",
-    "@angular/cdk": "~12.1.1",
-    "@angular/common": "~12.0.5",
-    "@angular/compiler": "~12.0.0",
-    "@angular/core": "~12.0.0",
-    "@angular/forms": "~12.0.5",
+    "@angular/animations": "~12.2.0",
+    "@angular/cdk": "~12.2.0",
+    "@angular/common": "~12.2.0",
+    "@angular/compiler": "~12.2.0",
+    "@angular/core": "~12.2.0",
+    "@angular/forms": "~12.2.0",
     "@angular/material": "^12.1.1",
-    "@angular/material-moment-adapter": "^12.1.1",
-    "@angular/platform-browser": "~12.0.0",
-    "@angular/platform-browser-dynamic": "~12.0.5",
-    "@angular/router": "~12.0.5",
-    "@angular/service-worker": "~12.0.5",
+    "@angular/material-moment-adapter": "^12.2.0",
+    "@angular/platform-browser": "~12.2.0",
+    "@angular/platform-browser-dynamic": "~12.2.0",
+    "@angular/router": "~12.2.0",
+    "@angular/service-worker": "~12.2.0",
     "@grpc/grpc-js": "^1.3.2",
     "@ngx-translate/core": "^13.0.0",
     "@ngx-translate/http-loader": "^6.0.0",
     "@types/file-saver": "^2.0.2",
-    "@types/google-protobuf": "^3.7.4",
+    "@types/google-protobuf": "^3.15.3",
     "@types/uuid": "^8.3.0",
     "angular-oauth2-oidc": "^10.0.3",
     "angularx-qrcode": "^11.0.0",
@@ -35,7 +35,7 @@
     "google-proto-files": "^2.4.0",
     "google-protobuf": "^3.17.2",
     "grpc-web": "^1.2.1",
-    "libphonenumber-js": "^1.9.16",
+    "libphonenumber-js": "^1.9.23",
     "moment": "^2.29.1",
     "ngx-color": "^7.2.0",
     "ngx-image-cropper": "^3.3.5",
@@ -49,11 +49,11 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~12.1.1",
-    "@angular/cli": "~12.1.1",
-    "@angular/compiler-cli": "~12.0.5",
-    "@angular/language-service": "~12.1.0",
-    "@types/jasmine": "~3.7.7",
-    "@types/jasminewd2": "~2.0.3",
+    "@angular/cli": "~12.2.0",
+    "@angular/compiler-cli": "~12.2.0",
+    "@angular/language-service": "~12.2.0",
+    "@types/jasmine": "~3.8.2",
+    "@types/jasminewd2": "~2.0.10",
     "@types/node": "^15.12.1",
     "codelyzer": "^6.0.0",
     "jasmine-core": "~3.7.1",
@@ -66,9 +66,9 @@
     "prettier": "^2.3.1",
     "protractor": "~7.0.0",
     "stylelint": "^13.10.0",
-    "stylelint-config-standard": "^20.0.0",
-    "stylelint-scss": "^3.19.0",
-    "ts-node": "~9.1.1",
+    "stylelint-config-standard": "^22.0.0",
+    "stylelint-scss": "^3.20.1",
+    "ts-node": "~10.2.0",
     "tslint": "~6.1.3",
     "typescript": "^4.2.4"
   }
